+++ anyway/widgets/suburban_widgets/most_severe_accidents_table_widget.py
@@ -3,7 +3,7 @@ from typing import Dict
 
 import pandas as pd
 from flask_babel import _
-from anyway.RequestParams import RequestParams
+from anyway.request_params import RequestParams
 from anyway.backend_constants import BE_CONST, AccidentSeverity, AccidentType
 from anyway.infographics_dictionaries import segment_dictionary
 from anyway.widgets.widget_utils import get_query, get_accidents_stats
