+++ anyway/widgets/suburban_widgets/most_severe_accidents_widget.py
@@ -3,7 +3,7 @@ from typing import Dict
 
 from flask_babel import _
 
-from anyway.RequestParams import RequestParams
+from anyway.request_params import RequestParams
 from anyway.backend_constants import AccidentSeverity, AccidentType
 from anyway.widgets.suburban_widgets.most_severe_accidents_table_widget import (
     get_most_severe_accidents_with_entities,
