+++ anyway/widgets/suburban_widgets/street_view_widget.py
@@ -1,4 +1,4 @@
-from anyway.RequestParams import RequestParams
+from anyway.request_params import RequestParams
 from anyway.widgets.widget import register
 from anyway.widgets.suburban_widgets.sub_urban_widget import SubUrbanWidget
