+++ anyway/widgets/urban_widgets/small_motor_severe_fatal_count_by_year_widget.py
@@ -1,6 +1,6 @@
 from typing import Dict
 
-from anyway.RequestParams import RequestParams
+from anyway.request_params import RequestParams
 from anyway.backend_constants import InjurySeverity
 from anyway.models import InvolvedMarkerView
 from anyway.vehicle_type import VehicleCategory
